+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added dispenser report download
 - added Crossmint, ENS, WalletConnect as a wallet options
 - bug fix for signer when checking ownership of tokens
+- added Invite Links page
 
 ## 2.1.22
 - allow pause and unpause dispenser